FlowCard as separate component
