+++ editor.planx.uk/src/pages/Team.tsx
@@ -1,19 +1,10 @@
-import { gql } from "@apollo/client";
-import MoreHoriz from "@mui/icons-material/MoreHoriz";
 import Box from "@mui/material/Box";
-import Button from "@mui/material/Button";
 import Container from "@mui/material/Container";
-import Dialog from "@mui/material/Dialog";
-import DialogActions from "@mui/material/DialogActions";
-import DialogContent from "@mui/material/DialogContent";
-import DialogContentText from "@mui/material/DialogContentText";
-import DialogTitle from "@mui/material/DialogTitle";
 import { styled } from "@mui/material/styles";
 import Typography from "@mui/material/Typography";
 import { hasFeatureFlag } from "lib/featureFlags";
 import React, { useCallback, useEffect, useState } from "react";
-import { Link, useNavigation } from "react-navi";
-import { inputFocusStyle } from "theme";
+import { useNavigation } from "react-navi";
 import { AddButton } from "ui/editor/AddButton";
 import { SortableFields, SortControl } from "ui/editor/SortControl";
 import Input from "ui/shared/Input/Input";
@@ -22,11 +13,9 @@ import InputRowItem from "ui/shared/InputRowItem";
 import InputRowLabel from "ui/shared/InputRowLabel";
 import { slugify } from "utils";
 
-import { client } from "../lib/graphql";
-import SimpleMenu from "../ui/editor/SimpleMenu";
+import FlowCard, { Card, CardContent } from "./FlowCard";
 import { useStore } from "./FlowEditor/lib/store";
 import { FlowSummary } from "./FlowEditor/lib/store/editor";
-import { formatLastEditMessage } from "./FlowEditor/utils";
 
 const DashboardList = styled("ul")(({ theme }) => ({
   padding: theme.spacing(3, 0),
@@ -43,279 +32,15 @@ const DashboardList = styled("ul")(({ theme }) => ({
   },
 }));
 
-const FlowCard = styled("li")(({ theme }) => ({
-  listStyle: "none",
-  position: "relative",
-  display: "flex",
-  flexDirection: "column",
-  justifyContent: "stretch",
-  borderRadius: "3px",
-  backgroundColor: theme.palette.background.default,
-  border: `1px solid ${theme.palette.border.main}`,
-  boxShadow: "0 2px 4px 0 rgba(0, 0, 0, 0.1)",
-}));
-
-const FlowCardContent = styled(Box)(({ theme }) => ({
-  position: "relative",
-  height: "100%",
-  textDecoration: "none",
-  color: "currentColor",
-  padding: theme.spacing(2),
-  display: "flex",
-  flexDirection: "column",
-  alignItems: "flex-start",
-  gap: theme.spacing(1.5),
-  margin: 0,
-  width: "100%",
-}));
-
-const DashboardLink = styled(Link)(() => ({
-  position: "absolute",
-  left: 0,
-  top: 0,
-  width: "100%",
-  height: "100%",
-  zIndex: 1,
-  "&:focus": {
-    ...inputFocusStyle,
-  },
-}));
-
-const LinkSubText = styled(Box)(({ theme }) => ({
-  color: theme.palette.text.secondary,
-  fontWeight: "normal",
-  paddingTop: theme.spacing(0.75),
-}));
-
-const StyledSimpleMenu = styled(SimpleMenu)(({ theme }) => ({
-  display: "flex",
-  marginTop: "auto",
-  borderTop: `1px solid ${theme.palette.border.main}`,
-  backgroundColor: theme.palette.background.paper,
-  overflow: "hidden",
-  borderRadius: "0px 0px 4px 4px",
-  "& > button": {
-    padding: theme.spacing(0.25, 1),
-    width: "100%",
-    justifyContent: "flex-start",
-    "& > svg": {
-      display: "none",
-    },
-  },
-}));
-
-const Confirm = ({
-  title,
-  content,
-  submitLabel,
-  open,
-  onClose,
-  onConfirm,
-}: {
-  title: string;
-  content: string;
-  submitLabel: string;
-  open: boolean;
-  onClose: () => void;
-  onConfirm: () => void;
-}) => (
-  <Dialog
-    open={open}
-    onClose={() => {
-      onClose();
-    }}
-  >
-    <DialogTitle>{title}</DialogTitle>
-    <DialogContent>
-      <DialogContentText>{content}</DialogContentText>
-    </DialogContent>
-    <DialogActions>
-      <Button onClick={onClose} color="primary">
-        Cancel
-      </Button>
-      <Button onClick={onConfirm} color="primary">
-        {submitLabel}
-      </Button>
-    </DialogActions>
-  </Dialog>
-);
-
-interface FlowItemProps {
-  flow: FlowSummary;
-  flows: FlowSummary[];
-  teamId: number;
-  teamSlug: string;
-  refreshFlows: () => void;
-}
-
-const FlowItem: React.FC<FlowItemProps> = ({
-  flow,
-  flows,
-  teamId,
-  teamSlug,
-  refreshFlows,
-}) => {
-  const [deleting, setDeleting] = useState(false);
-  const handleDelete = () => {
-    useStore
-      .getState()
-      .deleteFlow(teamId, flow.slug)
-      .then(() => {
-        setDeleting(false);
-        refreshFlows();
-      });
-  };
-  const handleCopy = () => {
-    useStore
-      .getState()
-      .copyFlow(flow.id)
-      .then(() => {
-        refreshFlows();
-      });
-  };
-  const handleMove = (newTeam: string) => {
-    useStore
-      .getState()
-      .moveFlow(flow.id, newTeam)
-      .then(() => {
-        refreshFlows();
-      });
-  };
-
-  return (
-    <>
-      {deleting && (
-        <Confirm
-          title="Confirm Delete"
-          open={deleting}
-          content="Deleting a service cannot be reversed."
-          onClose={() => {
-            setDeleting(false);
-          }}
-          onConfirm={handleDelete}
-          submitLabel="Delete Service"
-        />
-      )}
-      <FlowCard>
-        <FlowCardContent>
-          <Box>
-            <Typography variant="h3" component="h2">
-              {flow.name}
-            </Typography>
-            <LinkSubText>
-              {formatLastEditMessage(
-                flow.operations[0].createdAt,
-                flow.operations[0]?.actor,
-              )}
-            </LinkSubText>
-          </Box>
-          <DashboardLink
-            aria-label={flow.name}
-            href={`./${flow.slug}`}
-            prefetch={false}
-          />
-        </FlowCardContent>
-        {useStore.getState().canUserEditTeam(teamSlug) && (
-          <StyledSimpleMenu
-            items={[
-              {
-                onClick: async () => {
-                  const newName = prompt("New name", flow.name);
-                  if (newName && newName !== flow.name) {
-                    const newSlug = slugify(newName);
-                    const duplicateFlowName = flows?.find(
-                      (flow: any) => flow.slug === newSlug,
-                    );
-                    if (!duplicateFlowName) {
-                      await client.mutate({
-                        mutation: gql`
-                          mutation UpdateFlowSlug(
-                            $teamId: Int
-                            $slug: String
-                            $newSlug: String
-                            $newName: String
-                          ) {
-                            update_flows(
-                              where: {
-                                team: { id: { _eq: $teamId } }
-                                slug: { _eq: $slug }
-                              }
-                              _set: { slug: $newSlug, name: $newName }
-                            ) {
-                              affected_rows
-                            }
-                          }
-                        `,
-                        variables: {
-                          teamId: teamId,
-                          slug: flow.slug,
-                          newSlug: newSlug,
-                          newName: newName,
-                        },
-                      });
-
-                      refreshFlows();
-                    } else if (duplicateFlowName) {
-                      alert(
-                        `The flow "${newName}" already exists. Enter a unique flow name to continue`,
-                      );
-                    }
-                  }
-                },
-                label: "Rename",
-              },
-              {
-                label: "Copy",
-                onClick: () => {
-                  handleCopy();
-                },
-              },
-              {
-                label: "Move",
-                onClick: () => {
-                  const newTeam = prompt("New team");
-                  if (newTeam) {
-                    if (slugify(newTeam) === teamSlug) {
-                      alert(
-                        `This flow already belongs to ${teamSlug}, skipping move`,
-                      );
-                    } else {
-                      handleMove(slugify(newTeam));
-                    }
-                  }
-                },
-              },
-              {
-                label: "Delete",
-                onClick: () => {
-                  setDeleting(true);
-                },
-                error: true,
-              },
-            ]}
-          >
-            <Box sx={{ display: "flex", alignItems: "center", gap: 0.2 }}>
-              <MoreHoriz sx={{ fontSize: "1.4em" }} />
-              <Typography variant="body2" fontSize="small">
-                <strong>Menu</strong>
-              </Typography>
-            </Box>
-          </StyledSimpleMenu>
-        )}
-      </FlowCard>
-    </>
-  );
-};
-
 const GetStarted: React.FC<{ flows: FlowSummary[] }> = ({ flows }) => (
   <DashboardList sx={{ paddingTop: 0 }}>
-    <FlowCard>
-      <FlowCardContent>
+    <Card>
+      <CardContent>
         <Typography variant="h3">No services found</Typography>
         <Typography>Get started by creating your first service</Typography>
         <AddFlowButton flows={flows} />
-      </FlowCardContent>
-    </FlowCard>
+      </CardContent>
+    </Card>
   </DashboardList>
 );
 
@@ -412,7 +137,6 @@ const Team: React.FC = () => {
             <Typography variant="h2" component="h1" pr={1}>
               Services
             </Typography>
-            {/* {canUserEditTeam(slug) ? <Edit /> : <Visibility />} */}
             {showAddFlowButton && <AddFlowButton flows={flows} />}
           </Box>
           <Box maxWidth={360}>
@@ -448,7 +172,7 @@ const Team: React.FC = () => {
           {teamHasFlows && (
             <DashboardList>
               {flows.map((flow) => (
-                <FlowItem
+                <FlowCard
                   flow={flow}
                   flows={flows}
                   key={flow.slug}
